refactor: core모듈 구조 수정 (#59)
