+++ module-core/src/main/java/com/back2basics/post/model/Post.java
@@ -1,7 +1,7 @@
 package com.back2basics.post.model;
 
 import com.back2basics.comment.model.Comment;
-import com.back2basics.service.post.dto.PostUpdateCommand;
+import com.back2basics.post.port.in.command.PostUpdateCommand;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
