+++ module-core/src/main/java/com/back2basics/post/port/in/UpdatePostUseCase.java
@@ -1,6 +1,6 @@
 package com.back2basics.post.port.in;
 
-import com.back2basics.service.post.dto.PostUpdateCommand;
+import com.back2basics.post.port.in.command.PostUpdateCommand;
 
 public interface UpdatePostUseCase {
