--- module-infra/src/main/java/com/back2basics/adapter/persistence/example/port/ExampleOutPort.java
@@ -1,5 +1,0 @@
-package com.back2basics.adapter.persistence.example.port;
-
-public interface ExampleOutPort {
-
-}
