--- module-infra/src/main/java/com/back2basics/adapter/persistence/example/port/ExampleRepository.java
@@ -1,7 +1,0 @@
-package com.back2basics.adapter.persistence.example.port;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ExampleRepository extends JpaRepository<ExampleEntity, Long> {
-
-}
